move crd test to e2e
